refactor: decompose 596-line test-agent route into modules (#7)
Before: src/app/api/test-agent/route.ts — 596 lines (god file)
After:  596 → 83 lines (thin orchestrator) + 4 focused modules

New modules in src/lib/agent-runner/:
  workspace.ts  (28 lines)  — buildWorkspaceContext(sessionId)
  history.ts    (13 lines)  — buildHistory(messages)
  agents.ts     (56 lines)  — createAgents(agentDefs, model, ctx)
  stream.ts    (390 lines)  — streamSolo/streamReflect/streamCollaborate

route.ts (83 lines) — thin orchestrator only.
Zero new TS errors. Same request/response contract preserved.

Closes #7
